# internal/docs/swagger.yaml
basePath: /
definitions:
...
  model.Source:
    properties:
      chunk_count:
        type: integer
      created_at:
        type: string
      filename:
        type: string
      id:
        type: string
      mime_type:
        type: string
      notebook_id:
        type: string
      rag_doc_id:
        description: doc_id from rag-anything
        type: string
      status:
        $ref: '#/definitions/model.SourceStatus'
      storage_key:
        description: e.g. S3 object key
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  model.SourceStatus:
    enum:
    - pending
    - processing
    - ready
    - failed
    type: string
    x-enum-varnames:
    - SourceStatusPending
    - SourceStatusProcessing
    - SourceStatusReady
    - SourceStatusFailed
host: localhost:8080
info:
  contact: {}
  description: RAG-backed notebook and conversation management service.
  title: Notebook Service API
  version: "1.0"
paths:
...
  /notebooks/{notebookID}/sources/:
    get:
      description: Returns all source documents in a notebook.
      parameters:
      - description: Notebook UUID
        in: path
        name: notebookID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Source'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List sources
      tags:
      - sources
    post:
      description: Accepts a multipart upload, creates a source record, and enqueues
        async ingestion into rag-anything.
      parameters:
      - description: Notebook UUID
        in: path
        name: notebookID
        required: true
        type: string
      - description: Document file (PDF, DOCX, etc.)
        in: formData
        name: file
        required: true
        type: file
      - description: Pre-uploaded object storage key (defaults to notebookID/filename)
        in: formData
        name: storage_key
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted — ingestion is asynchronous
          schema:
            $ref: '#/definitions/model.Source'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: File too large (max 50 MB)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload a source document
      tags:
      - sources
  /notebooks/{notebookID}/sources/{sourceID}:
    delete:
      description: Permanently deletes a source document and removes it from rag-anything.
      parameters:
      - description: Notebook UUID
        in: path
        name: notebookID
        required: true
        type: string
      - description: Source UUID
        in: path
        name: sourceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a source
      tags:
      - sources
    get:
      description: Returns a single source document including its ingestion status.
      parameters:
      - description: Notebook UUID
        in: path
        name: notebookID
        required: true
        type: string
      - description: Source UUID
        in: path
        name: sourceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Source'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a source
      tags:
      - sources
securityDefinitions:
  BearerAuth:
    description: JWT issued by Authentik. Prefix with "Bearer ".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"